pyproject.toml
[tool.poetry]
name = "nexus"
version = "0.1.0"
description = ""
authors = ["juanj72 <juan.jara@campusucc.edu.co>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.5"
mysqlclient = "^2.2.7"
python-dotenv = "^1.0.1"
gdal = "==3.6.*"
django-leaflet = "^0.31.0"
djangorestframework = "^3.15.2"
drf-yasg = "^1.21.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
